vita3k: code cleanup and minor bugfixes
Fix bugs
* gui/disassembly_dialog: fix bug. Can't enter addresses >= 0x80000000
* utils/vc_runtime_checker: fix almost impossible AV in vc_runtime_checker
* util/net_utils: fix warning and error in parseHeaders function
* patch: fix almost impossible bug, also drop unused var
* modules/SceAvPlayer: fix error (wrong mutex)
* kernel: more correct value of export variable __sce_libcparam (NID:0xDF084DFA)
* interface.cpp/get_contents_path (Looks like mistype in duplicate search)

Needs recheck of refactoring. I found something strange, but I'm not sure if it's intended or not.
* renderer::vulkan/VKContext::VKContext (Looks like mistype in variable name)
* spirv_recompiler/create_fragment_inputs (Variable anonimous is always false here. It's not static and reinit every loop iteration)
* renderer/transfer: fix strange (and maybe never run) code if pixel format is unknown
* renderer/vulkan/replacement: fix some type to type converson

Refactoring and enhancements
* codec/atrac9: better error logs
* util/fs_utils: Function read_data which read file to vector. Also use it if possible
* util/log: custom formatter for Ptr<> type (it simplifies logging)
* util/tracy: set tracy_module_id variable static
* util/warning.h : header to cross-platform disable warnings
* renderer/transfer: fix strange (and maybe never run) code if pixel format is unknown
* nids: typo fix
* modules/module_parent: remove link to nids.inc from header
* modules/module_parent: move macro to other such macros in bridge.h
* kernel: refactoring use get_thread instead of lock_and_find
* mem/ptr: set type Ptr<T> trivial (trivially copiable, movable and so on)

Fix warnings:
* some retype in BitmapAllocator (mem-tests are OK)
* typecasts
* remove unused variables and fields
* push_back->emplace_back
* rename duplicate variables (usually in nested loops)
* drop excessive brackets
* remove some unused includes
* set 'static' for functions if possible
* gui:home_screen/draw_home_screen: same values around ? operator.
* convert declare SCE_*_ERROR_* to enums
* bracket initialisation instead of memset
* other compiler warnings

+++ vita3k/codec/src/h264.cpp
@@ -26,26 +26,26 @@ extern "C" {
 #include <cassert>
 
 void copy_yuv_data_from_frame(AVFrame *frame, uint8_t *dest, const uint32_t width, const uint32_t height, bool is_p3) {
-    for (uint32_t i = 0; i < height; i++) {
+    for (size_t i = 0; i < height; i++) {
         memcpy(dest, &frame->data[0][frame->linesize[0] * i], width);
         dest += width;
     }
 
     if (is_p3) {
-        for (uint32_t i = 0; i < height / 2; i++) {
+        for (size_t i = 0; i < height / 2; i++) {
             memcpy(dest, &frame->data[1][frame->linesize[1] * i], width / 2);
             dest += width / 2;
         }
-        for (uint32_t i = 0; i < height / 2; i++) {
+        for (size_t i = 0; i < height / 2; i++) {
             memcpy(dest, &frame->data[2][frame->linesize[2] * i], width / 2);
             dest += width / 2;
         }
     } else {
         // p2 format, U and V are interleaved
-        for (uint32_t i = 0; i < height / 2; i++) {
+        for (size_t i = 0; i < height / 2; i++) {
             const uint8_t *src_u = &frame->data[1][frame->linesize[1] * i];
             const uint8_t *src_v = &frame->data[2][frame->linesize[2] * i];
-            for (uint32_t j = 0; j < width / 2; j++) {
+            for (size_t j = 0; j < width / 2; j++) {
                 dest[0] = src_u[j];
                 dest[1] = src_v[j];
                 dest += 2;
